Revert "tidy up code"
