+++ luasrc/model/cbi/koolproxy/rss_rule.lua
@@ -2,8 +2,7 @@ local m, s, o
 local koolproxy = "koolproxy"
 local sid = arg[1]
 
-m = Map(koolproxy)
-m.title = translate("GodProxy滤广告 - 编辑规则")
+m = Map(koolproxy, "%s - %s" %{translate("GodProxy滤广告"), translate("编辑规则")})
 m.redirect = luci.dispatcher.build_url("admin/services/koolproxy")
 
 if not arg[1] or m.uci:get(koolproxy, sid) ~= "rss_rule" then
@@ -16,15 +15,15 @@ s = m:section(NamedSection, sid, "rss_rule")
 s.anonymous = true
 s.addremove = true
 
-o = s:option(Flag, "load", translate("启用"))
-o.rmempty = false
-o.default = 0
+o=s:option(Flag,"load",translate("启用"))
+o.default=0
+o.rmempty=false
 
-o = s:option(Value, "name", translate("规则描述"))
-o.rmempty = true
+o=s:option(Value,"name",translate("规则描述"))
+o.rmempty=true
 
-o = s:option(Value, "url", translate("规则地址"))
-o.rmempty = false
+o=s:option(Value,"url",translate("规则地址"))
+o.rmempty=false
 o.placeholder="[https|http|ftp]://[Hostname]/[File]"
 function o.validate(self, value)
 	if not value then
