--- luasrc/model/cbi/koolproxy/tips.lua
@@ -1,8 +1,0 @@
-o = Map("koolproxy")
-
-t = o:section(TypedSection, "usetips")
-t.title = translate("GodProxy 帮助支持")
-t.anonymous = true
-t:append(Template("koolproxy/tips"))
-
-return o
