+++ luasrc/view/koolproxy/caupload.htm
@@ -1,5 +1,5 @@
 <%+cbi/valueheader%>
-    <label class="cbi-value" style="display:inline-block; width: 400px" for="ulfile"></label><br />
+    <label class="cbi-value" style="display:inline-block; width: 400px" for="ulfile"><font color="red"><%:Upload backup file,The file name must be koolproxyCA.tar.gz%></font></label><br />
 	<input class="cbi-input-file" style="width: 400px" type="file" id="ulfile" name="ulfile" />
-	<input type="submit" class="cbi-button cbi-input-apply" name="upload" value="<%:上传恢复证书%>" />
+	<input type="submit" class="cbi-button cbi-input-apply" name="upload" value="<%:Upload Restore%>" />
 <%+cbi/valuefooter%>
